ui связан с odata v4 стратегии и бэктест
