+++ static/webapp/view/App.view.xml
@@ -1,5 +1,4 @@
-<mvc:View controllerName="fokin.crypto.controller.App"
-	xmlns:mvc="sap.ui.core.mvc"
+<mvc:View xmlns:mvc="sap.ui.core.mvc"
 	displayBlock="true"
 	xmlns="sap.m">
 	<App id="app"/>
